requirements.txt
pandas>=1.5.0
numpy>=1.21.0
ta>=0.10.0  # Technical analysis library
python-binance>=1.0.16  # For Binance API integration
ccxt>=2.0.0  # For multi-exchange support
python-dotenv>=0.19.0  # For environment variables
pytest>=7.0.0  # For testing 